Remove FormBuilderSubHeader
