--- frontend/form-builder/src/components/FormBuilderSubHeader/FormBuilderSubHeader.module.css
@@ -1,39 +1,0 @@
-.subHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: var(--fds-sizing-10);
-  background: var(--fds-semantic-surface-neutral-inverted);
-  color: var(--fds-semantic-surface-on_inverted-default);
-  padding: 0 var(--fds-spacing-12);
-
-  a,
-  a:hover {
-    text-decoration: none;
-    color: inherit !important;
-  }
-}
-
-.subHeaderHeading {
-  color: var(--fds-semantic-surface-on_inverted-default);
-}
-
-.subHeaderLinks {
-  display: flex;
-  gap: var(--fds-spacing-10);
-}
-
-.subHeaderIcon {
-  font-size: 1.5rem;
-}
-
-@media screen and (max-width: 600px) {
-  .subHeader {
-    padding: 0 var(--fds-spacing-4);
-  }
-
-  .subHeaderLinks {
-    gap: var(--fds-spacing-6);
-  }
-}
