--- frontend/form-builder/src/components/FormBuilderSubHeader/index.ts
@@ -1,1 +1,0 @@
-export { FormBuilderSubHeader } from "./FormBuilderSubHeader";
